pyproject.toml
[project]
name = "meeting-capture"
version = "0.1.0"
description = "Always-on local audio capture daemon — system audio to transcripts on disk, fully offline."
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "soundfile>=0.12",
    "mlx-whisper>=0.4.0; sys_platform == 'darwin'",
    "pyobjc-framework-AVFoundation>=10.0; sys_platform == 'darwin'",
]

[project.scripts]
meeting-capture = "meeting_capture.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meeting_capture"]
